whereami counter

Interact with internal whereami counter

counter_state(item = NULL, tag = NULL)

counter_reset(item = NULL, tag = NULL)

counter_names()

counter_tags()

counter_get()

Arguments

item
character, name of the counter to access, Default: NULL

tag
character, tag of the counter to acces, if it was set at initial whereami call. This can be used instead of the item name, Default: NULL

Details

counter_names() returns names of the active stored counters.

counter_state() returns current hit count for item, if NULL then all counters are returned.

counter_reset() will remove counter of item, if item is NULL then all counters are reset.

counter_get() returns a table with all the counters. These can then be plotted with the autoplot() method.

Examples

if( interactive() ){

tf <- tempfile(fileext = '.R')

# call to write to temp file
f <- "whereami::cat_where(whereami::whereami(tag = 'my tag'))"

# write to the file (call at line 3)
cat('\n\n',f,sep='',file = tf)

source(tf)

counter_state()

counter_state(tag = 'my tag')

counter_names()

# using counter_reset and counter_names in a loop

for( i in 1:10 ){

  source(tf)

  if( counter_state(counter_names()[1]) > 5 )
    counter_reset(counter_names()[1])
}

# clear all counters
counter_reset()

# cleanup
unlink(tf)

}